+++ sources/color_effects.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 16:44:29 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/02/18 17:36:24 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/02/20 02:17:47 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@ int	escape_time_color(t_fractol *ctl, int iterations)
 
 int	lerp_color(t_fractol *ctl, int iterations)
 {
-	double	normalized_iterations;
-	double	intensity;
-	int		color;
+	long double	normalized_iterations;
+	long double	intensity;
+	int			color;
 
 	normalized_iterations = ft_map_d((t_map_d){iterations, 0,
 			ctl->max_iterations, 0, 1});
@@ -37,7 +37,7 @@ int	bernstein_color(t_fractol *ctl, int iterations)
 	unsigned char	red;
 	unsigned char	green;
 	unsigned char	blue;
-	double			ni;
+	long double		ni;
 
 	ni = ft_map_d((t_map_d){iterations, 0, ctl->max_iterations, 0, 1});
 	red = 9.0 * (1.0 - ni) * pow(ni, 3.0) * 255.0;
